+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fix cursor position after Extract Variable that creates a shorthand on an object property that ends with a comma, so rename works
+- Stop proposing Negate Expression quick fix on single call expressions, since the negated expression is identical
 
 ## [1.2.0] - 2019-11-28 - Big head, shorthand 🦖